requirements.txt
easyocr==1.7.2
gdown==5.2.0
nltk==3.9.1
numpy==2.1.3
oauthlib==3.2.2
opencv-python-headless==4.10.0.84
opentelemetry-api==1.27.0
ordered-set==4.0.2
packaging==24.1
pandas==2.2.3
Pillow==11.0.0
PyMySQL==1.1.1
python_bcrypt==0.3.2
spacy==3.7.5
streamlit==1.40.0
streamlit_camera_input_live==0.2.0
streamlit_extras==0.5.0
torch==2.5.1
tqdm==4.67.0
scikit-learn